+++ angular/test/base/scripts/sync.sh
@@ -16,3 +16,6 @@ npm pack ../../../../packages/angular-server/dist
 
 # Install Dependencies
 npm install *.tgz --no-save
+
+# Delete Angular cache directory
+rm -rf .angular/