+++ angular/test/base/src/app/overlays-inline/overlays-inline.component.ts
@@ -1,0 +1,31 @@
+import { Component } from "@angular/core";
+
+/**
+ * Validates that inline overlays will will display correctly.
+ */
+@Component({
+  selector: 'app-overlays-inline',
+  templateUrl: 'overlays-inline.component.html'
+})
+export class OverlaysInlineComponent {
+  public pickerButtons = [{ text: 'Ok' }, { text: 'Cancel', role: 'cancel' }];
+  public pickerColumns = [
+    {
+      name: 'Colors',
+      options: [
+        {
+          text: 'Red',
+          value: 'red',
+        },
+        {
+          text: 'Blue',
+          value: 'blue',
+        },
+        {
+          text: 'Green',
+          value: 'green',
+        },
+      ],
+    },
+  ];
+}