+++ core/src/components/datetime/utils/manipulation.ts
@@ -37,18 +37,6 @@ export function convertDataToISO(data: DatetimeParts | DatetimeParts[]): string 
         if (data.hour !== undefined) {
           // YYYY-MM-DDTHH:mm:SS
           rtn += `T${twoDigit(data.hour)}:${twoDigit(data.minute)}:00`;
-
-          if (data.tzOffset === undefined) {
-            // YYYY-MM-DDTHH:mm:SSZ
-            rtn += 'Z';
-          } else {
-            // YYYY-MM-DDTHH:mm:SS+/-HH:mm
-            rtn +=
-              (data.tzOffset > 0 ? '+' : '-') +
-              twoDigit(Math.floor(Math.abs(data.tzOffset / 60))) +
-              ':' +
-              twoDigit(data.tzOffset % 60);
-          }
         }
       }
     }