+++ core/src/components/searchbar/searchbar.tsx
@@ -4,10 +4,12 @@ import { arrowBackSharp, closeCircle, closeSharp, searchOutline, searchSharp } f
 
 import { config } from '../../global/config';
 import { getIonMode } from '../../global/ionic-global';
-import type { AutocompleteTypes, Color, SearchbarChangeEventDetail, StyleEventDetail } from '../../interface';
+import type { AutocompleteTypes, Color, StyleEventDetail } from '../../interface';
 import { debounceEvent, raf } from '../../utils/helpers';
 import { isRTL } from '../../utils/rtl';
 import { createColorClasses } from '../../utils/theme';
+
+import type { SearchbarChangeEventDetail, SearchbarInputEventDetail } from './searchbar-interface';
 
 /**
  * @virtualProp {"ios" | "md"} mode - The mode determines which platform styles to use.
@@ -24,6 +26,12 @@ export class Searchbar implements ComponentInterface {
   private nativeInput?: HTMLInputElement;
   private isCancelVisible = false;
   private shouldAlignLeft = true;
+  private originalIonInput?: EventEmitter<SearchbarInputEventDetail>;
+
+  /**
+   * The value of the input when the textarea is focused.
+   */
+  private focusedValue?: string | null;
 
   @Element() el!: HTMLIonSearchbarElement;
 
@@ -69,13 +77,19 @@ export class Searchbar implements ComponentInterface {
   @Prop() clearIcon?: string;
 
   /**
-   * Set the amount of time, in milliseconds, to wait to trigger the `ionChange` event after each keystroke. This also impacts form bindings such as `ngModel` or `v-model`.
+   * Set the amount of time, in milliseconds, to wait to trigger the `ionInput` event after each keystroke.
    */
-  @Prop() debounce = 250;
+  @Prop() debounce?: number;
 
   @Watch('debounce')
   protected debounceChanged() {
-    this.ionChange = debounceEvent(this.ionChange, this.debounce);
+    const { ionInput, debounce, originalIonInput } = this;
+
+    /**
+     * If debounce is undefined, we have to manually revert the ionInput emitter in case
+     * debounce used to be set to a number. Otherwise, the event would stay debounced.
+     */
+    this.ionInput = debounce === undefined ? originalIonInput ?? ionInput : debounceEvent(ionInput, debounce);
   }
 
   /**
@@ -149,12 +163,19 @@ export class Searchbar implements ComponentInterface {
   @Prop({ mutable: true }) value?: string | null = '';
 
   /**
-   * Emitted when a keyboard input occurred.
+   * Emitted when the `value` of the `ion-searchbar` element has changed.
    */
-  @Event() ionInput!: EventEmitter<KeyboardEvent>;
+  @Event() ionInput!: EventEmitter<SearchbarInputEventDetail>;
 
   /**
-   * Emitted when the value has changed.
+   * The `ionChange` event is fired for `<ion-searchbar>` elements when the user
+   * modifies the element's value. Unlike the `ionInput` event, the `ionChange`
+   * event is not necessarily fired for each alteration to an element's value.
+   *
+   * The `ionChange` event is fired when the value has been committed
+   * by the user. This can happen when the element loses focus or
+   * when the "Enter" key is pressed. `ionChange` can also fire
+   * when clicking the clear or cancel buttons.
    */
   @Event() ionChange!: EventEmitter<SearchbarChangeEventDetail>;
 
@@ -191,7 +212,6 @@ export class Searchbar implements ComponentInterface {
     if (inputEl && inputEl.value !== value) {
       inputEl.value = value;
     }
-    this.ionChange.emit({ value });
   }
 
   @Watch('showCancelButton')
@@ -207,6 +227,7 @@ export class Searchbar implements ComponentInterface {
   }
 
   componentDidLoad() {
+    this.originalIonInput = this.ionInput;
     this.positionElements();
     this.debounceChanged();
 
@@ -241,30 +262,65 @@ export class Searchbar implements ComponentInterface {
   }
 
   /**
+   * Emits an `ionChange` event.
+   *
+   * This API should be called for user committed changes.
+   * This API should not be used for external value changes.
+   */
+  private emitValueChange(event?: Event) {
+    const { value } = this;
+    // Checks for both null and undefined values
+    const newValue = value == null ? value : value.toString();
+    // Emitting a value change should update the internal state for tracking the focused value
+    this.focusedValue = newValue;
+    this.ionChange.emit({ value: newValue, event });
+  }
+
+  /**
+   * Emits an `ionInput` event.
+   */
+  private emitInputChange(event?: Event) {
+    const { value } = this;
+    this.ionInput.emit({ value, event });
+  }
+
+  /**
    * Clears the input field and triggers the control change.
    */
-  private onClearInput = (shouldFocus?: boolean) => {
+  private onClearInput = async (shouldFocus?: boolean) => {
     this.ionClear.emit();
 
-    // setTimeout() fixes https://github.com/ionic-team/ionic/issues/7527
-    // wait for 4 frames
-    setTimeout(() => {
-      const value = this.getValue();
-      if (value !== '') {
-        this.value = '';
-        this.ionInput.emit();
+    return new Promise<void>((resolve) => {
+      // setTimeout() fixes https://github.com/ionic-team/ionic/issues/7527
+      // wait for 4 frames
+      setTimeout(() => {
+        const value = this.getValue();
+        if (value !== '') {
+          this.value = '';
+          this.emitInputChange();
 
-        /**
-         * When tapping clear button
-         * ensure input is focused after
-         * clearing input so users
-         * can quickly start typing.
-         */
-        if (shouldFocus && !this.focused) {
-          this.setFocus();
+          /**
+           * When tapping clear button
+           * ensure input is focused after
+           * clearing input so users
+           * can quickly start typing.
+           */
+          if (shouldFocus && !this.focused) {
+            this.setFocus();
+
+            /**
+             * The setFocus call above will clear focusedValue,
+             * but ionChange will never have gotten a chance to
+             * fire. Manually revert focusedValue so onBlur can
+             * compare against what was in the box before the clear.
+             */
+            this.focusedValue = value;
+          }
         }
-      }
-    }, 16 * 4);
+
+        resolve();
+      }, 16 * 4);
+    });
   };
 
   /**
@@ -272,13 +328,27 @@ export class Searchbar implements ComponentInterface {
    * the clearInput function doesn't want the input to blur
    * then calls the custom cancel function if the user passed one in.
    */
-  private onCancelSearchbar = (ev?: Event) => {
+  private onCancelSearchbar = async (ev?: Event) => {
     if (ev) {
       ev.preventDefault();
       ev.stopPropagation();
     }
     this.ionCancel.emit();
-    this.onClearInput();
+
+    // get cached values before clearing the input
+    const value = this.getValue();
+    const focused = this.focused;
+
+    await this.onClearInput();
+
+    /**
+     * If there used to be something in the box, and we weren't focused
+     * beforehand (meaning no blur fired that would already handle this),
+     * manually fire ionChange.
+     */
+    if (value && !focused) {
+      this.emitValueChange(ev);
+    }
 
     if (this.nativeInput) {
       this.nativeInput.blur();
@@ -288,22 +358,31 @@ export class Searchbar implements ComponentInterface {
   /**
    * Update the Searchbar input value when the input changes
    */
-  private onInput = (ev: Event) => {
+  private onInput = (ev: InputEvent | Event) => {
     const input = ev.target as HTMLInputElement | null;
     if (input) {
       this.value = input.value;
     }
-    this.ionInput.emit(ev as KeyboardEvent);
+    this.emitInputChange(ev);
+  };
+
+  private onChange = (ev: Event) => {
+    this.emitValueChange(ev);
   };
 
   /**
    * Sets the Searchbar to not focused and checks if it should align left
    * based on whether there is a value in the searchbar or not.
    */
-  private onBlur = () => {
+  private onBlur = (ev: FocusEvent) => {
     this.focused = false;
     this.ionBlur.emit();
     this.positionElements();
+
+    if (this.focusedValue !== this.value) {
+      this.emitValueChange(ev);
+    }
+    this.focusedValue = undefined;
   };
 
   /**
@@ -311,6 +390,7 @@ export class Searchbar implements ComponentInterface {
    */
   private onFocus = () => {
     this.focused = true;
+    this.focusedValue = this.value;
     this.ionFocus.emit();
     this.positionElements();
   };
@@ -503,6 +583,7 @@ export class Searchbar implements ComponentInterface {
             inputMode={this.inputmode}
             enterKeyHint={this.enterkeyhint}
             onInput={this.onInput}
+            onChange={this.onChange}
             onBlur={this.onBlur}
             onFocus={this.onFocus}
             placeholder={this.placeholder}