+++ core/src/components/slides/swiper/swiper.js
@@ -1,4 +1,0 @@
-import { Autoplay, Pagination, Scrollbar, Swiper, Keyboard, Zoom } from 'swiper/js/swiper.esm';
-
-Swiper.use([Pagination, Scrollbar, Autoplay, Keyboard, Zoom]);
-export { Swiper };