+++ core/src/utils/config.ts
@@ -1,4 +1,6 @@
-import type { AnimationBuilder, Mode, SpinnerTypes, TabButtonLayout } from '../interface';
+import type { SpinnerTypes } from '../components/spinner/spinner-configs';
+import type { TabButtonLayout } from '../components/tab-bar/tab-bar-interface';
+import type { AnimationBuilder, Mode } from '../interface';
 
 import type { PlatformConfig } from './platform';
 
@@ -191,7 +193,9 @@ export interface IonicConfig {
    * Relevant Components: ion-alert, ion-infinite-scroll-content, ion-loading, ion-refresher-content, ion-toast
    * If `false`, all `innerHTML` usage will be disabled in Ionic, and
    * custom HTML will not be usable in the relevant components.
-   * `innerHTML` usage is enabled by default.
+   * If `true`, all `innerHTML` usage will be enabled in Ionic, and
+   * custom HTML will be usable in the relevant components.
+   * `innerHTML` usage is disabled by default.
    */
   innerHTMLTemplatesEnabled?: boolean;
 
@@ -247,4 +251,4 @@ export const getMode = (): Mode => {
   return 'md';
 };
 
-export const ENABLE_HTML_CONTENT_DEFAULT = true;
+export const ENABLE_HTML_CONTENT_DEFAULT = false;