+++ core/src/utils/theme.ts
@@ -1,4 +1,5 @@
-import type { AnimationBuilder, Color, CssClassMap, RouterDirection } from '../interface';
+import type { RouterDirection } from '../components/router/utils/interface';
+import type { AnimationBuilder, Color, CssClassMap } from '../interface';
 
 export const hostContext = (selector: string, el: HTMLElement): boolean => {
   return el.closest(selector) !== null;