+++ core/stencil.config.ts
@@ -9,6 +9,7 @@ import { apiSpecGenerator } from './scripts/api-spec-generator';
 
 export const config: Config = {
   autoprefixCss: true,
+  sourceMap: false,
   namespace: 'Ionic',
   bundles: [
     { components: ['ion-action-sheet'] },
@@ -45,7 +46,6 @@ export const config: Config = {
     { components: ['ion-searchbar'] },
     { components: ['ion-segment', 'ion-segment-button'] },
     { components: ['ion-select', 'ion-select-option', 'ion-select-popover'] },
-    { components: ['ion-slides', 'ion-slide'] },
     { components: ['ion-spinner'] },
     { components: ['ion-split-pane'] },
     { components: ['ion-tabs', 'ion-tab'] },
@@ -53,7 +53,6 @@ export const config: Config = {
     { components: ['ion-text'] },
     { components: ['ion-toast'] },
     { components: ['ion-toggle'] },
-    { components: ['ion-virtual-scroll'] },
     { components: ['ion-accordion-group', 'ion-accordion'] },
     { components: ['ion-breadcrumb', 'ion-breadcrumbs'] },
   ],
@@ -137,10 +136,16 @@ export const config: Config = {
           externalEvent: 'ionChange'
         },
         {
-          elements: ['ion-datetime', 'ion-input', 'ion-radio-group', 'ion-radio', 'ion-range', 'ion-searchbar', 'ion-segment', 'ion-segment-button', 'ion-select', 'ion-textarea', 'ion-accordion-group'],
+          elements: ['ion-datetime', 'ion-radio-group', 'ion-radio', 'ion-range', 'ion-segment', 'ion-segment-button', 'ion-select', 'ion-accordion-group'],
           targetAttr: 'value',
           event: 'v-ion-change',
           externalEvent: 'ionChange'
+        },
+        {
+          elements: ['ion-input', 'ion-searchbar', 'ion-textarea'],
+          targetAttr: 'value',
+          event: 'v-ion-input',
+          externalEvent: 'ionInput'
         }
       ],
     }),
@@ -182,15 +187,9 @@ export const config: Config = {
       directivesProxyFile: '../angular/src/directives/proxies.ts',
       directivesArrayFile: '../angular/src/directives/proxies-list.ts',
       excludeComponents: [
-        // overlays
-        'ion-action-sheet',
-        'ion-alert',
-        'ion-loading',
+        // overlays that accept user components
         'ion-modal',
-        'ion-picker',
         'ion-popover',
-        'ion-toast',
-        'ion-toast',
 
         // navigation
         'ion-router',
@@ -205,16 +204,10 @@ export const config: Config = {
 
         // auxiliar
         'ion-picker-column',
-        'ion-virtual-scroll'
       ],
     }),
   ],
   buildEs5: 'prod',
-  extras: {
-    dynamicImportShim: true,
-    initializeNextTick: true,
-    scriptDataOpts: true
-  },
   testing: {
     moduleNameMapper: {
       "@utils/test": ["<rootDir>/src/utils/test/utils"],
@@ -224,4 +217,5 @@ export const config: Config = {
   preamble: '(C) Ionic http://ionicframework.com - MIT License',
   globalScript: 'src/global/ionic-global.ts',
   enableCache: true,
+  transformAliasedImportPaths: true,
 };