+++ packages/vue/test/base/src/views/Home.vue
@@ -26,9 +26,6 @@
         <ion-item button router-link="/inputs">
           <ion-label>Inputs</ion-label>
         </ion-item>
-        <ion-item button router-link="/slides">
-          <ion-label>Slides</ion-label>
-        </ion-item>
         <ion-item button router-link="/navigation" id="navigation">
           <ion-label>Navigation</ion-label>
         </ion-item>