+++ packages/vue/test/base/tests/unit/hooks.spec.ts
@@ -1,4 +1,5 @@
 import { mount } from '@vue/test-utils';
+import { describe, expect, it } from 'vitest';
 import { createRouter, createWebHistory } from '@ionic/vue-router';
 import { IonicVue, IonApp, IonRouterOutlet, IonPage, useIonRouter, createAnimation } from '@ionic/vue';
 import { mockAnimation, waitForRouter } from './utils';
@@ -14,15 +15,6 @@ const BasePage = {
 }
 
 describe('useIonRouter', () => {
-  beforeAll(() => {
-    (HTMLElement.prototype as HTMLIonRouterOutletElement).commit = jest.fn((entering, leaving, opts) => {
-      if (opts && opts.animationBuilder) {
-        opts.animationBuilder(entering, leaving);
-      }
-
-      return Promise.resolve(true);
-    });
-  });
   it('should correctly navigate back', async () => {
     const Page1 = {
       ...BasePage